[store-review] Transform StoreReview into unimodule
