--- ios/Exponent/Versioned/Core/Api/EXStoreReview.h
@@ -1,14 +1,0 @@
-//
-//  EXStoreReview.h
-//  Exponent
-//
-//  Created by Evan Bacon on 4/17/18.
-//  Copyright © 2018 650 Industries. All rights reserved.
-//
-
-#import <React/RCTBridgeModule.h>
-#import <Foundation/Foundation.h>
-
-@interface EXStoreReview : NSObject <RCTBridgeModule>
-
-@end
