+++ packages/expo-store-review/.npmignore
@@ -1,0 +1,2 @@
+# @generated by expo-module-scripts
+babel.config.js
