+++ packages/expo-store-review/README.md
@@ -1,0 +1,30 @@
+# expo-store-review
+
+`expo-store-review` allows users rate your application inside App Store.
+
+# API documentation
+
+- [Documentation for the master branch](https://github.com/expo/expo/blob/master/docs/pages/versions/unversioned/sdk/storereview.md)
+- [Documentation for the latest stable release](https://docs.expo.io/versions/latest/sdk/storereview/)
+
+# Installation
+
+You must ensure that you have [installed and configured the `react-native-unimodules` package](https://github.com/unimodules/react-native-unimodules) before continuing.
+
+### Add the package to your dependencies
+
+```
+npm install expo-store-review
+```
+
+### Configure for iOS
+
+Run `pod install` in the `ios` directory after installing the npm package.
+
+### Configure for Android
+
+No additional set up necessary.
+
+# Contributing
+
+Contributions are very welcome! Please refer to guidelines described in the [contributing guide]( https://github.com/expo/expo#contributing).
