+++ packages/expo-store-review/babel.config.js
@@ -1,0 +1,2 @@
+// @generated by expo-module-scripts
+module.exports = require('expo-module-scripts/babel.config.base');
