+++ packages/expo-store-review/ios/EXStoreReview/EXStoreReviewModule.h
@@ -1,0 +1,8 @@
+//  Copyright © 2018 650 Industries. All rights reserved.
+
+#import <UMCore/UMExportedModule.h>
+#import <UMCore/UMModuleRegistryConsumer.h>
+
+@interface EXStoreReviewModule : UMExportedModule
+
+@end
