+++ packages/expo-store-review/src/ExpoStoreReview.ts
@@ -1,0 +1,2 @@
+import { NativeModulesProxy } from '@unimodules/core';
+export default NativeModulesProxy.ExpoStoreReview;
