+++ packages/expo-store-review/src/ExpoStoreReview.web.ts
@@ -1,0 +1,8 @@
+export default {
+  get name(): string {
+    return 'ExpoStoreReview';
+  },
+  get isSupported(): boolean {
+    return false;
+  }
+};
