+++ packages/expo-store-review/unimodule.json
@@ -1,0 +1,4 @@
+{
+  "name": "expo-store-review",
+  "platforms": ["ios"]
+}
