--- packages/expo/src/StoreReview/ExponentStoreReview.ts
@@ -1,2 +1,0 @@
-import { NativeModules } from 'react-native';
-export default NativeModules.ExponentStoreReview;
