--- packages/expo/src/StoreReview/ExponentStoreReview.web.ts
@@ -1,8 +1,0 @@
-export default {
-  get name(): string {
-    return 'ExponentStoreReview';
-  },
-  get isSupported(): boolean {
-    return false;
-  }
-};
